rearrange schema transaction tests
